requirements.txt
line-bot-sdk==3.5.0
pigpio==1.78
pyaudio==0.2.14
pygame==2.5.2
matplotlib==3.8.0
numpy==1.26.1
boto3==1.34.2
slackweb==1.0.5
fastapi== 0.109.0
uvicorn==0.25.0
boto3==1.34.2
uvicorn==0.25.0
python-dotenv==1.0.1
firebase-admin==6.5.0